fix(avatarcropper): demo拆解与规范
