--- src/packages/avatarcropper/demo.scss
@@ -1,5 +1,0 @@
-.toolbar {
-  display: flex;
-  justify-content: space-between;
-  padding: 10px;
-}
